Use CsvExporter helper class to reduce low level code on account bloc
